Add files via upload
Creating Rock, Paper, and Scissors game with Python. In this game my goal was to create a command-line game where a user has the option to choose between rock, paper and scissors and if the user wins the score gets addition assigned, and at the end when the user finishes the game by typing 'End', the score is shown to the user.
I have tested and implemented Random module and nested If-Else statements.